+++ Source/ImGui/Private/ImGuiInputHandlerFactory.h
@@ -2,6 +2,9 @@
 
 #pragma once
 
+#include "ImGuiModuleSettings.h"
+
+
 class FImGuiModuleManager;
 class UGameViewportClient;
 class UImGuiInputHandler;